refactor(problem): 페이지 계산 기능 모듈 분리
