+++ common/mathrank-page/src/test/java/kr/co/mathrank/PageApplication.java
@@ -1,0 +1,7 @@
+package kr.co.mathrank;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PageApplication {
+}
