+++ domain/mathrank-problem-domain/build.gradle
@@ -12,6 +12,7 @@ dependencies {
 
     implementation project(':common:mathrank-snowflake')
     implementation project(':common:mathrank-exception')
+    implementation project(':common:mathrank-page')
     api project(':domain:mathrank-problem-core-domain')
 
     implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
